[FLINK-36009] Move ContinuousUnBoundingSettings to source
