+++ flink-connector-jdbc-core/src/main/java/org/apache/flink/connector/jdbc/core/datastream/source/config/ContinuousUnBoundingSettings.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.connector.jdbc.utils;
+package org.apache.flink.connector.jdbc.core.datastream.source.config;
 
 import org.apache.flink.annotation.PublicEvolving;
 
@@ -31,7 +31,7 @@ import static org.apache.flink.util.Preconditions.checkNotNull;
  * continuous discovery and streaming mode.
  */
 @PublicEvolving
-public final class ContinuousUnBoundingSettings implements Serializable {
+public class ContinuousUnBoundingSettings implements Serializable {
 
     private static final long serialVersionUID = 1L;
